% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_gam_formula.R
\name{auto_gam_formula}
\alias{auto_gam_formula}
\title{Construct a GAM formula from a tsibble}
\usage{
auto_gam_formula(
  data,
  response,
  box_cox = FALSE,
  box_cox_lambda = "auto",
  xregs = NULL,
  trend = TRUE,
  trend_args = list(),
  season = TRUE,
  season_spec = NULL,
  threshold = 24,
  K_d = 2
)
}
\arguments{
\item{data}{A tsibble containing a single measured variable.}

\item{response}{The name of the response variable.}

\item{box_cox}{Whether to apply a box_cox transformation to the response.}

\item{box_cox_lambda}{A numeric value for the box_cox transformation parameter}

\item{xregs}{A character vector containing regressor variables' names.}

\item{trend}{Does the model include trend?}

\item{trend_args}{A list of named arguments to pass to `trend()`}

\item{season}{Does the model include seasonalities?}

\item{season_spec}{Optional manual specification of seasonalities}

\item{threshold}{Maximum seasonal periods that triggers `season()` instead of `fourier()`}

\item{K_d}{Denominator of the maximum number of trigonometric terms \eqn{K = \lfloor period/K_d \rfloor`}}
}
\value{
A model formula suitable for [`GAM()`].
}
\description{
This helper inspects the series index using [`fabletools::common_periods()`]
to include appropriate `trend()`, `season()` or `fourier()` specials.
}
\details{
`season()` terms are used for short integer periods (\eqn{\le threshold}), while
longer or non integer periods are expressed with `fourier()` using
\eqn{K = \lfloor period/K_d \rfloor`}.
}
