% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{detect_outliers}
\alias{detect_outliers}
\title{Detect Outliers for a GAM model}
\usage{
detect_outliers(.mable, .tsibble, level = 95, ...)
}
\arguments{
\item{.mable}{A mable containing the fitted GAM.}

\item{.tsibble}{A tsibble containing the original series}

\item{level}{confidence level for detecting the outliers}

\item{...}{Further arguments for methods.}
}
\description{
The table enables to select the observations which appears to be outliers.
}
