% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exact_periods}
\alias{exact_periods}
\title{Extract exact frequencies for common seasonal periods}
\usage{
exact_periods(x)
}
\arguments{
\item{x}{A tsibble.}
}
\value{
A named vector of frequencies appropriate for the provided data.
}
\description{
This helper is derived from [`fabletools::common_periods()`]
but without the approximation of weekly data.
}
