% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{package}
\name{fable.GAM-package}
\alias{fable.GAM}
\alias{fable.GAM-package}
\title{fable.GAM: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MauroFalc1/fable.GAM}
  \item Report bugs at \url{https://github.com/MauroFalc1/fable.GAM/issues}
}

}
\author{
\strong{Maintainer}: Jane Doe \email{jane@example.com}

}
\keyword{package}
