% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.fbl_gam}
\alias{fitted.fbl_gam}
\title{Extract fitted values}
\usage{
\method{fitted}{fbl_gam}(object, ...)
}
\arguments{
\item{object}{A model for which forecasts are required.}

\item{...}{Other arguments passed to methods}
}
\value{
Vector of fitted values
}
\description{
Extracts the fitted values from an estimated GAM
}
\author{
Trent Henderson
}
