% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{forecast.fbl_gam}
\alias{forecast.fbl_gam}
\title{Produce forecasts from the GAM}
\usage{
\method{forecast}{fbl_gam}(object, new_data, specials = NULL, ...)
}
\arguments{
\item{object}{A model for which forecasts are required.}

\item{new_data}{A tsibble containing the time points and exogenous regressors to produce forecasts for.}

\item{specials}{(passed by \code{\link[fabletools:forecast]{fabletools::forecast.mdl_df()}}).}

\item{...}{Additional arguments passed to [`stats::predict()`].}
}
\value{
A list of forecasts
}
\description{
If additional future information is required, such as exogenous variables,
then they should be included as variables of the `new_data` argument.
}
\examples{

\donttest{
if (requireNamespace("tsibble")) {
library(tsibble)
tsibble::tourism \%>\%
  filter(Region == "Melbourne") \%>\%
  model(mygam = GAM(Trips ~ trend() + season(4))) \%>\%
  forecast(h = "5 years")
}
}

}
\author{
Trent Henderson
}
