% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{forecast.GAM}
\alias{forecast.GAM}
\title{Produce forecasts from a GAM model}
\usage{
\method{forecast}{GAM}(
  object,
  new_data,
  specials = NULL,
  bootstrap = FALSE,
  approx_normal = TRUE,
  times = 5000,
  ...
)
}
\arguments{
\item{object}{A model for which forecasts are required.}

\item{new_data}{A tsibble containing the time points and exogenous regressors to produce forecasts for.}

\item{specials}{(passed by \code{\link[fabletools:forecast]{fabletools::forecast.mdl_df()}}).}

\item{bootstrap}{If \code{TRUE}, then forecast distributions are computed using simulation with resampled errors.}

\item{approx_normal}{Should Gaussian forecasts be approximated
by a Normal (default) rather than a Student‑t?}

\item{times}{The number of sample paths to use in estimating the forecast distribution when \code{bootstrap = TRUE}.}

\item{...}{Other arguments passed to methods}
}
\value{
A list of forecasts.
}
\description{
Produces forecasts from a trained model as in the `fable` package.
}
\examples{
USAccDeaths \%>\%
  as_tsibble() \%>\%
  model(gam = GAM(log(value) ~ trend() + season())) \%>\%
  forecast()
}
