% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{generate.fbl_gam}
\alias{generate.fbl_gam}
\title{Generate future sample paths from a GAM}
\usage{
\method{generate}{fbl_gam}(x, new_data, specials = NULL, ...)
}
\arguments{
\item{x}{\code{fbl_gam} model object}

\item{new_data}{\code{tsibble} of future time points at which to generate paths}

\item{specials}{Parsed specials from the model formula}

\item{...}{arguments to be passed to methods}
}
\description{
Generates simulated future sample paths from a fitted `GAM`. Innovations are
resampled from the in-sample residuals, giving a non-parametric empirical bootstrap.
}
\author{
Trent Henderson
}
