% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{generate.GAM}
\alias{generate.GAM}
\title{Generate future sample paths from a <GAM> model}
\usage{
generate.GAM(x, new_data, specials = NULL, bootstrap = FALSE, times = 1, ...)
}
\arguments{
\item{x}{A fitted model.}

\item{new_data}{A tsibble containing the time points and exogenous regressors to produce forecasts for.}

\item{specials}{(passed by \code{\link[fabletools:forecast]{fabletools::forecast.mdl_df()}}).}

\item{bootstrap}{If \code{TRUE}, then forecast distributions are computed using simulation with resampled errors.}

\item{times}{Integer. Number of simulated paths to produce. Default 1.}

\item{...}{Other arguments passed to methods}
}
\description{
Generate future sample paths from a <GAM> model
}
