% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{glance.fbl_gam}
\alias{glance.fbl_gam}
\title{Glance a GAM}
\usage{
\method{glance}{fbl_gam}(x, ...)
}
\arguments{
\item{x}{model or other R object to convert to single-row data frame}

\item{...}{other arguments passed to methods}
}
\value{
`tibble` summarising the model fit.
}
\description{
Construct a single row summary of the GAM model. Contains a range of model fit statistics.
}
\examples{
\donttest{
library(dplyr)
library(tsibble)

tourism_melb <- tsibble::tourism |>
  filter(Region == "Melbourne") |>
  filter(Purpose == "Business") |>
  dplyr::select(c(Quarter, Region, Trips)) |>
  as_tsibble(key = Region, index = Quarter)

fit <- tourism_melb |>
  model(mygam = GAM(Trips ~ trend() + season(4))) |>
  glance()
}
}
