% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{interpolate.fbl_gam}
\alias{interpolate.fbl_gam}
\title{Interpolate missing values using the GAM}
\usage{
\method{interpolate}{fbl_gam}(object, new_data, specials, ...)
}
\arguments{
\item{object}{\code{fbl_gam} model object}

\item{new_data}{\code{tsibble} containing observations, some of which may be \code{NA}}

\item{specials}{Parsed specials from the model formula.}

\item{...}{arguments to be passed to methods}
}
\value{
The \code{new_data} tsibble with missing values replaced by GAM predictions.
}
\description{
Replaces \code{NA} values in the response variable with in-sample GAM
predictions evaluated at those time points.
}
\author{
Trent Henderson
}
