% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{plot_outliers}
\alias{plot_outliers}
\title{Detect Outliers for a GAM model}
\usage{
plot_outliers(.mable, .tsibble, level = 95, ...)
}
\arguments{
\item{.mable}{A mable containing the fitted GAM.}

\item{.tsibble}{A tsibble containing the original series}

\item{level}{confidence level for detecting the outliers}

\item{...}{Further arguments for methods.}
}
\description{
The plots show the component contributions, on the link scale, of each model term to the linear predictor.
}
