% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{plot_partial_effects}
\alias{plot_partial_effects}
\title{Partial Effect Plots for GAM}
\usage{
plot_partial_effects(x, se = TRUE, ...)
}
\arguments{
\item{x}{A mable object.}

\item{se}{if \code{TRUE}, upper and lower pointwise
twice-standard-error curves are included for each plot. The
default is \code{FALSE}.}

\item{...}{Further arguments for methods.}
}
\description{
The plots show the component contributions, on the link scale, of each model term to the linear predictor.
}
\seealso{
[`gam::plot.Gam()`]
}
