% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{report.GAM}
\alias{report.GAM}
\title{Summarise a GAM model fit (Hastie & Tibshirani `gam` package)}
\usage{
\method{report}{GAM}(object, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{A fitted GAM object (class `GAM` created by `train_gam()` or
equivalent helper inside this package).}

\item{digits}{Minimum number of significant digits to print.}

\item{...}{Further arguments passed for S3 completeness (currently
ignored).}
}
\value{
(Invisibly) returns `object` so the call can be piped/assigned.
}
\description{
Provides a console report formatted similarly to other `fable` model
`report()` methods, but tailored for objects fitted with the **gam** package.
In particular it makes use of the `anova` and `parametric.anova` tables
returned by `summary.gam()`.
}
\examples{
as_tsibble(USAccDeaths) \%>\%
  model(gam = GAM(log(value) ~ trend() + season())) \%>\%
  report()
}
