% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.fbl_gam}
\alias{residuals.fbl_gam}
\title{Extract model residuals}
\usage{
\method{residuals}{fbl_gam}(object, ...)
}
\arguments{
\item{object}{A model for which forecasts are required.}

\item{...}{Other arguments passed to methods}
}
\value{
Vector of residuals
}
\description{
Extracts the residuals from an estimated GAM
}
\author{
Trent Henderson
}
