% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{residuals.GAM}
\alias{residuals.GAM}
\title{Extract residuals from a fable model}
\usage{
\method{residuals}{GAM}(
  object,
  type = c("innovation", "response", "deviance", "pearson", "working", "partial"),
  ...
)
}
\arguments{
\item{object}{A model for which forecasts are required.}

\item{type}{The type of residuals to extract.}

\item{...}{Other arguments passed to methods}
}
\value{
A vector of fitted residuals.
}
\description{
Extracts the residuals.
}
\examples{
as_tsibble(USAccDeaths) \%>\%
  model(gam = GAM(log(value) ~ trend() + season())) \%>\%
  residuals()
}
