% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{specials_gam}
\alias{specials_gam}
\title{Special functions for the GAM model}
\usage{
specials_gam
}
\description{
These special functions provide interfaces to more complicated functions
within the model formulae interface.
}
\section{Specials}{


\subsection{trend}{
The `trend` special includes common linear trend regressors in the model.
It also supports piecewise linear trend via the `knots` argument.
\preformatted{
trend(knots = NULL, origin = NULL, linear=FALSE, ...)
}

\tabular{ll}{
  `knots`    \tab A vector of times (same class as the data's time index)
  identifying the position of knots for a piecewise linear trend.\cr
  `origin`   \tab An optional starting time value for the trend. \cr
  `linear` \tab logical indicating whether using linear or smoothed trend.\cr
  `...` \tab other variables to be passed to `gam::s()`
}
}

\subsection{season}{
The `season` special includes seasonal dummy variables in the model.
\preformatted{
season(period = NULL)
}

\tabular{ll}{
  `period`   \tab The periodic nature of the seasonality. This can be either a number indicating the number of observations in each seasonal period, or text to indicate the duration of the seasonal window (for example, annual seasonality would be "1 year").
}
}

\subsection{fourier}{
The `fourier` special includes seasonal fourier terms in the model. The maximum order of the fourier terms must be specified using `K`.
\preformatted{
fourier(period = NULL, K, origin = NULL)
}

\tabular{ll}{
  `period`   \tab The periodic nature of the seasonality. This can be either a number indicating the number of observations in each seasonal period, or text to indicate the duration of the seasonal window (for example, annual seasonality would be "1 year"). \cr
  `K`        \tab The maximum order of the fourier terms.\cr
  `origin`   \tab An optional time value to act as the starting time for the fourier series.
}
}
}

\seealso{
* [fabletools::new_specials()],
* <https://fabletools.tidyverts.org/articles/extension_models.html>
}
\keyword{internal}
