% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tidy.fbl_gam}
\alias{tidy.fbl_gam}
\title{Present a tidy summary of a GAM}
\usage{
\method{tidy}{fbl_gam}(x, ...)
}
\description{
Present a tidy summary of a GAM
}
\examples{
if (requireNamespace("tsibble")) {
library(tsibble)
tsibble::tourism \%>\%
  filter(Region == "Melbourne") \%>\%
  model(mygam = GAM(Trips ~ trend() + season(4))) \%>\%
  tidy()
}
}
