% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{train_gam}
\alias{train_gam}
\title{Training function of GAM model}
\usage{
train_gam(.data, specials, ...)
}
\arguments{
\item{.data}{tsibble containing the transformed response}

\item{specials}{list of specials defined for GAM}

\item{...}{other variables to be passed to `gam::gam()`}
}
\value{
a model object of class GAM
}
\description{
`train_gam()` is the internal engine that powers the user‑facing [GAM()]
model specification. It prepares the data, translates fable specials into
smooth terms, and calls [gam::gam()] from the `gam` package to estimate a
univariate additive model.
}
\seealso{
[gam::gam()],
<https://www.statlearning.com/>
}
\keyword{internal}
