% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{components.GAM}
\alias{components.GAM}
\title{Extract components from a GAM model}
\usage{
\method{components}{GAM}(object, ...)
}
\arguments{
\item{object}{A fitted model of class GAM.}

\item{...}{Currently ignored.}
}
\value{
A [`fabletools::dable`] whose columns sum to `.response` *exactly*.
}
\description{
Extract components from a GAM model
}
\examples{
library(fabletools)
as_tsibble(USAccDeaths) \%>\%
  model(gam = GAM(log(value) ~ trend() + season())) \%>\%
  components()
}
