% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{package}
\name{fable.GAM-package}
\alias{fable.GAM}
\alias{fable.GAM-package}
\title{fable.GAM: Generalized Additive Model (GAM) wrapper for fable}
\description{
Use GAMs within the fable/fabletools workframe.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MauroFalc1/fable.GAM}
  \item Report bugs at \url{https://github.com/MauroFalc1/fable.GAM/issues}
}

}
\author{
\strong{Maintainer}: Mauro Falcone \email{mauro.falcone@live.it}

}
\keyword{package}
