% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{forecast.fbl_gam}
\alias{forecast.fbl_gam}
\title{Produce forecasts from the GAM}
\usage{
\method{forecast}{fbl_gam}(object, new_data, specials = NULL, ...)
}
\arguments{
\item{object}{A model for which forecasts are required.}

\item{new_data}{A tsibble containing the time points and exogenous regressors to produce forecasts for.}

\item{specials}{(passed by \code{\link[fabletools:forecast]{fabletools::forecast.mdl_df()}}).}

\item{...}{arguments passed to \code{stats::predict()}.}
}
\value{
A vector of distributions.
}
\description{
If additional future information is required, such as exogenous variables, then they should be included as variables of the \code{new_data} argument.
}
\examples{

\donttest{
tourism_melb <- tsibble::tourism |>
  filter(Region == "Melbourne") |>
  filter(Purpose == "Business") |>
  dplyr::select(c(Quarter, Region, Trips)) |>
  as_tsibble(key = Region, index = Quarter)

fit <- tourism_melb |>
  model(mygam = GAM(Trips ~ trend() + season(4)))

# Analytic intervals

forecast(fit, h = "5 years")

# Bootstrap intervals

forecast(fit, h = "5 years", bootstrap = TRUE)
}

}
\author{
Trent Henderson
}
