% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_gam.R
\name{GAM}
\alias{GAM}
\title{Generalised additive modelling}
\usage{
GAM(formula, ...)
}
\arguments{
\item{formula}{A symbolic description of the model to be fitted of class `formula`.}

\item{...}{further arguments for 
passing on e.g. to \code{gam.fit} (such as \code{mustart}). }
}
\description{
Prepares a generalised additive model specification for use within the `fable` package.
}
\details{
The GAM modelling interface uses a `formula` based model specification
`y ~ x`, where the left of the formula specifies the response variable,
and the right specifies the model's predictive terms, including any smooth
functions of exogenous regressors or 'covariates'.
}
\section{Specials}{


\subsection{trend}{
The `trend` special is used to specify a seasonal component. This special can be used multiple times for different seasonalities.

\preformatted{
trend()
}
}

\subsection{season}{
The `season` special is used to specify a seasonal component. This special can be used multiple times for different seasonalities.

**NOTE: The inputs controlling the seasonal `period` refer to the number of observations in each seasonal period, not the number of days.**

\preformatted{
season(period = NULL)
}
}

\subsection{xreg}{
Exogenous regressors can be passed into `GAM` just like a regular `mgcv::gam` call by specifying the variable name or any smooth terms you want to model.
}
}

\examples{
if (requireNamespace("tsibble")) {
library(tsibble)
tsibble::tourism \%>\%
  filter(Region == "Melbourne") \%>\%
  model(mygam = GAM(Trips ~ trend() + season(4)))
}

}
\author{
Trent Henderson
}
