% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{GAM}
\alias{GAM}
\title{Fit a Generalized Additive Model (GAM) with time series components}
\usage{
GAM(formula, ...)
}
\arguments{
\item{formula}{Model specification.}
}
\value{
A model specification.
}
\description{
The model formula will be handled using [`gam::gam()`], and so
the same approach to include smoothing functions in [`gam::gam()`] applies when
specifying the `formula`. In addition, it is possible to
include [`specials_gam`] in the model formula, such as `trend()`, `season()`,
and `fourier()`.
}
\section{Specials}{


\subsection{xreg}{
Exogenous regressors can be included in a GAM model without explicitly using the `xreg()` special.
Specials as specified in [`specials_gam`] can also be used.
\preformatted{
xreg(...)
}

\tabular{ll}{
  `...`      \tab Bare expressions for the exogenous regressors (such as `s(x)`)
}
}
}

\examples{
as_tsibble(USAccDeaths) \%>\%
  model(gam = GAM(log(value) ~ trend() + season()))

library(tsibbledata)
olympic_running \%>\%
  model(GAM(Time ~ trend())) \%>\%
  interpolate(olympic_running)
}
\seealso{
[`gam::gam()`],
[An Introduction to Statistical Learning, Moving Beyond Linearity (chapter 7)](https://www.statlearning.com/)
}
