% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{GAM}
\alias{GAM}
\title{Fit a Generalized Additive Model (GAM) with time series components}
\usage{
GAM(formula, ...)
}
\arguments{
\item{formula}{Model specification.}

\item{...}{Any other orguments being passed to [`gam::gam()`]}
}
\value{
A model specification.
}
\description{
The model formula will be handled using [`gam::gam()`], and so
the same approach to include smoothing functions in [`gam::gam()`] applies when
specifying the `formula`. In addition, it is possible to
include [`specials_gam`] in the model formula, such as `trend()`, `season()`,
and `fourier()`.
}
\section{Specials}{


\subsection{trend}{
The `trend` special includes common linear and smoothened trend regressors in the model.
It also supports piecewise linear trend via the `knots` argument.
\preformatted{
trend(knots = NULL, origin = NULL, linear=FALSE, ...)
}

\tabular{ll}{
  `knots`    \tab A vector of times (same class as the data's time index)
  identifying the position of knots for a piecewise linear trend.\cr
  `origin`   \tab An optional starting time value for the trend. \cr
  `linear` \tab logical indicating whether using linear or smoothened trend.\cr
  `...` \tab other variables to be passed to `gam::s()`
}
}

\subsection{season}{
The `season` special includes seasonal dummy variables in the model.
\preformatted{
season(period = NULL)
}

\tabular{ll}{
  `period`   \tab The periodic nature of the seasonality. This can be either a number indicating the number of observations in each seasonal period, or text to indicate the duration of the seasonal window (for example, annual seasonality would be "1 year").
}
}

\subsection{fourier}{
The `fourier` special includes seasonal fourier terms in the model. The maximum order of the fourier terms must be specified using `K`.
\preformatted{
fourier(period = NULL, K, origin = NULL)
}

\tabular{ll}{
  `period`   \tab The periodic nature of the seasonality. This can be either a number indicating the number of observations in each seasonal period, or text to indicate the duration of the seasonal window (for example, annual seasonality would be "1 year"). \cr
  `K`        \tab The maximum order of the fourier terms.\cr
  `origin`   \tab An optional time value to act as the starting time for the fourier series.
}
}

\subsection{xreg}{
Exogenous regressors can be included in a GAM model without explicitly using the `xreg()` special.
\preformatted{
xreg(...)
}

\tabular{ll}{
  `...`      \tab Bare expressions for the exogenous regressors (such as `s(x)`)
}
}
}

\examples{
library(fabletools)
as_tsibble(USAccDeaths) \%>\%
  model(gam = GAM(log(value) ~ trend() + season()))

library(tsibbledata)
olympic_running \%>\%
  model(GAM(Time ~ trend())) \%>\%
  interpolate(olympic_running)
}
\seealso{
[`gam::gam()`],
[An Introduction to Statistical Learning, Moving Beyond Linearity (chapter 7)](https://www.statlearning.com/)
}
