% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{plot_outliers}
\alias{plot_outliers}
\title{Plot Outliers from a GAM model}
\usage{
plot_outliers(
  .outliers,
  fill = "brown",
  alpha = 0.1,
  out_col = "blue",
  out_size = 3,
  act_col = "white",
  ...
)
}
\arguments{
\item{.outliers}{The resulting tsibble of the function `detect_outliers()`}

\item{fill}{Colour of the confidence interval of the fitted values}

\item{alpha}{Transparency of the confidence interval}

\item{out_col}{Colour of the outliers}

\item{out_size}{Size of the outliers}

\item{act_col}{Coulour of the actual observations}

\item{...}{Further arguments for methods.}
}
\description{
The plot shows the observation which appears to be outliers.
}
