% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{refit.GAM}
\alias{refit.GAM}
\title{Objects exported from other packages}
\usage{
refit.GAM(object, new_data, specials = NULL, ...)
}
\value{
A newly fitted object of class <GAM> ready to slot back into a mable.
}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{distributional}{\code{\link[distributional]{hilo}}}

  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}, \code{\link[dplyr]{vars}}}

  \item{generics}{\code{\link[generics]{accuracy}}, \code{\link[generics]{augment}}, \code{\link[generics]{components}}, \code{\link[generics]{equation}}, \code{\link[generics]{forecast}}, \code{\link[generics]{generate}}, \code{\link[generics]{glance}}, \code{\link[generics]{hypothesize}}, \code{\link[generics]{interpolate}}, \code{\link[generics]{refit}}, \code{\link[generics]{tidy}}}

  \item{ggplot2}{\code{\link[ggplot2]{autolayer}}, \code{\link[ggplot2]{autoplot}}}

  \item{tsibble}{\code{\link[tsibble:as-tsibble]{as_tsibble}}}
}
}
