% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tidy.fbl_gam}
\alias{tidy.fbl_gam}
\title{Present a tidy summary of a GAM}
\usage{
\method{tidy}{fbl_gam}(x, ...)
}
\arguments{
\item{x}{\code{fbl_gam} object}

\item{...}{arguments to be passed to methods}
}
\description{
Present a tidy summary of a GAM
}
\examples{
\donttest{
library(dplyr)
library(tsibble)

tourism_melb <- tsibble::tourism |>
  filter(Region == "Melbourne") |>
  filter(Purpose == "Business") |>
  dplyr::select(c(Quarter, Region, Trips)) |>
  as_tsibble(key = Region, index = Quarter)

fit <- tourism_melb |>
  model(mygam = GAM(Trips ~ trend() + season(4))) |>
  tidy()
}
}
